Fix bug in vrr_mod.c:id_show.
Printf format for the vrr id was "%d". This produced negative values
in /sys/kernel/vrr/id when the id overflowed into negative. Since
the id is unsigned, it should always printed positive. So changed
format to "%u\n". Added the newline in order to be consistent with
similar sysfs files.